move pure functions to helpers.js
